rename `masks.py` -> ancillary.py` and update docstring
