+++ seaice_ecdr/daily_aggregate.py
@@ -12,9 +12,9 @@ from loguru import logger
 from pm_tb_data._types import Hemisphere
 
 from seaice_ecdr._types import ECDR_SUPPORTED_RESOLUTIONS
+from seaice_ecdr.ancillary import get_ancillary_ds
 from seaice_ecdr.complete_daily_ecdr import get_ecdr_filepath
 from seaice_ecdr.constants import STANDARD_BASE_OUTPUT_DIR
-from seaice_ecdr.masks import get_ancillary_ds
 from seaice_ecdr.nc_attrs import get_global_attrs
 from seaice_ecdr.util import sat_from_filename, standard_daily_aggregate_filename
